Fixes an off by one error in chrome.
Chrome was reporting the value el.scrollTop + el.clientHeight as one less than reality. This adds 3 to the value (just to be sure) and checks for greater than or equal and returns true in a greater range. This has the added benefit of anticipating the intentions of the user if they don't quite make it to the exact bottom of the element. (I may actually raise this value to exploit this further.)